+++ pkg/scheduler/queue/tree_queue_test.go
@@ -10,6 +10,9 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+// TODO (casie): When deprecating TreeQueue, it's safe to delete this file wholesale; all relevant test
+//  functionality has been duplicated for MultiQueuingAlgorithmTreeQueue.
+
 // TestDequeueBalancedTree checks dequeuing behavior from a balanced tree.
 //
 // Dequeuing from a balanced tree allows the test to have a simple looped structures
@@ -28,7 +31,7 @@ func TestDequeueBalancedTree(t *testing.T) {
 	dequeuedPathCache := make([]QueuePath, rotationsBeforeRepeat)
 
 	for !root.IsEmpty() {
-		v := root.Dequeue()
+		_, v := root.Dequeue()
 		dequeuedPath := getChildPathFromQueueItem(v)
 
 		// require dequeued path has not repeated before the expected number of rotations
@@ -96,7 +99,7 @@ func TestDequeueUnbalancedTree(t *testing.T) {
 	root := makeUnbalancedTreeQueue(t)
 
 	// dequeue from root until exhausted
-	v := root.Dequeue()
+	_, v := root.Dequeue()
 	require.Equal(t, "root:0:val0", v)
 
 	// root:0 and any subtrees are exhausted
@@ -106,25 +109,25 @@ func TestDequeueUnbalancedTree(t *testing.T) {
 	// root:0 was deleted
 	require.Nil(t, root.getNode(path))
 
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	require.Equal(t, "root:1:val0", v)
 
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	require.Equal(t, "root:2:0:val0", v)
 
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	require.Equal(t, "root:1:0:val0", v)
 
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	require.Equal(t, "root:2:1:val0", v)
 
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	require.Equal(t, "root:1:val1", v)
 
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	require.Equal(t, "root:2:0:val1", v)
 
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	require.Equal(t, "root:1:0:val1", v)
 
 	// root:1 and any subtrees are exhausted
@@ -134,10 +137,10 @@ func TestDequeueUnbalancedTree(t *testing.T) {
 	// root:1 was deleted
 	require.Nil(t, root.getNode(path))
 
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	require.Equal(t, "root:2:1:val1", v)
 
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	require.Equal(t, "root:2:1:val2", v)
 
 	// root:2 and any subtrees are exhausted
@@ -264,19 +267,19 @@ func TestEnqueueDuringDequeueRespectsRoundRobin(t *testing.T) {
 	require.Equal(t, []string{"0", "1", "2"}, root.childQueueOrder)
 
 	// dequeue first item
-	v := root.Dequeue()
+	_, v := root.Dequeue()
 	dequeuedPath := getChildPathFromQueueItem(v)
 	require.Equal(t, QueuePath{"0"}, dequeuedPath)
 
 	// dequeue second item; root:1 is now exhausted and deleted
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	dequeuedPath = getChildPathFromQueueItem(v)
 	require.Equal(t, QueuePath{"1"}, dequeuedPath)
 	require.Nil(t, root.getNode(QueuePath{"1"}))
 	require.Equal(t, []string{"0", "2"}, root.childQueueOrder)
 
 	// dequeue third item
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	dequeuedPath = getChildPathFromQueueItem(v)
 	require.Equal(t, QueuePath{"2"}, dequeuedPath)
 
@@ -290,18 +293,18 @@ func TestEnqueueDuringDequeueRespectsRoundRobin(t *testing.T) {
 
 	// dequeue fourth item; the newly-enqueued root:1 item
 	// has not jumped the line in front of root:0
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	dequeuedPath = getChildPathFromQueueItem(v)
 	require.Equal(t, QueuePath{"0"}, dequeuedPath)
 
 	// dequeue fifth item; the newly-enqueued root:1 item
 	// has not jumped the line in front of root:2
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	dequeuedPath = getChildPathFromQueueItem(v)
 	require.Equal(t, QueuePath{"2"}, dequeuedPath)
 
 	// dequeue sixth item; verifying the order 0->2->1 is being followed
-	v = root.Dequeue()
+	_, v = root.Dequeue()
 	dequeuedPath = getChildPathFromQueueItem(v)
 	require.Equal(t, QueuePath{"1"}, dequeuedPath)